+++ renderer/elements/icon/index.js
