+++ renderer/elements/player/styles.js
@@ -1,0 +1,8 @@
+var css = require('csjs-inject')
+
+module.exports = css`
+  .player {
+    display: flex;
+    align-items: center;
+  }
+`
