+++ renderer/elements/search/styles.js
@@ -1,0 +1,12 @@
+var css = require('csjs-inject')
+
+module.exports = css`
+  .searchInput {
+    width: auto;
+    padding: 1px 5px;
+    vertical-align: middle;
+    min-height: auto;
+    margin: 0 5px;
+    height: 24px;
+  }
+`
