+++ renderer/elements/table/index.js
@@ -1,0 +1,77 @@
+var html = require('choo/html')
+var fd = require('format-duration')
+var styles = require('./styles')
+
+function table (state, prev, send) {
+  return html`
+    <section class="${styles.pane}">
+      <div class=${styles.tableHeader}>
+        <table class="${styles.mediaList}">
+          <thead>
+            <tr>
+              <th>Title</th>
+              <th class="${styles.time}">Time</th>
+              <th>Artist</th>
+              <th>Album</th>
+            </tr>
+          </thead>
+        </table>
+      </div>
+      <div class=${styles.tableBody}>
+        <table class="${styles.mediaList} ${styles.tableBody}">
+          <tbody>${renderList(state, send)}</tbody>
+        </table>
+      </div>
+    </section>
+  `
+}
+
+function renderList (state, send) {
+  var { files, search } = state.library
+  var list = sortList(files)
+
+  if (search) list = filterList(list, search)
+
+  return list.map(meta => {
+    return html`
+      <tr onclick=${(e) => send('player:play', meta)}>
+        <td>${meta.title}</td>
+        <td class="${styles.time}">${meta.duration ? fd(meta.duration * 1000) : ''}</td>
+        <td>${meta.artist}</td>
+        <td>${meta.album}</td>
+      </tr>
+    `
+  })
+}
+
+// TODO: expose sort to state to allow sort using column headers
+function sortList (files) {
+  return files.sort((a, b) => {
+    // sort by artist
+    if (a.artist < b.artist) return -1
+    if (a.artist > b.artist) return 1
+
+    // then by album
+    if (a.album < b.album) return -1
+    if (a.album > b.album) return 1
+
+    // then by title
+    if (a.title < b.title) return -1
+    if (a.title > b.title) return 1
+    return 0
+  })
+}
+
+function filterList (list, search) {
+  return list.filter(meta => {
+    var yep = Object.keys(meta)
+      .map(i => (meta[i] + '').toLowerCase())
+      .filter(s => s.includes(search.toLowerCase()))
+      .length > 0
+
+    if (yep) return meta
+    return false
+  })
+}
+
+module.exports = table
