+++ renderer/elements/volume/styles.js
@@ -1,0 +1,12 @@
+var css = require('csjs-inject')
+
+module.exports = css`
+  .volumeControl {
+    position: relative;
+    display: inline-block;
+    vertical-align: middle;
+    -webkit-app-region: no-drag;
+    width: 100px;
+    margin: 0 5px;
+  }
+`
