+++ renderer/pages/main/index.js
@@ -1,7 +1,7 @@
 var html = require('choo/html')
-var toolbar = require('./components/toolbar')
-var table = require('./components/table')
-var styles = require('./styles')
+var toolbar = require('../../elements/toolbar')
+var table = require('../../elements/table')
+var styles = require('../styles')
 
 module.exports = (state, prev, send) => html`
   <main class="${styles.window}">
